pyproject.toml
[project]
name = "retailflow-ai"
version = "0.1.0"
description = "RetailFlow AI — B2B Smart Supply, Billing & Demand Intelligence System"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-nvidia-ai-endpoints>=0.1.0",
    "langchain-text-splitters>=0.2.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=3.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "passlib[bcrypt]>=1.7.4",
    "slowapi>=0.1.9",
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "mcp[cli]>=1.0.0",
    "streamlit>=1.56.0",
    "plotly>=6.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project.scripts]
init-db      = "scripts.init_db:main"
seed-db      = "scripts.seed_db:main"
init-rag     = "scripts.init_rag:main"
retailflow-mcp = "mcp_server:mcp.run"
